+++ mutwo/csound_converters/configurations.py
@@ -1,5 +1,8 @@
 """Configure the behaviour of :mod:`mutwo.csound_converters`."""
 
+CSOUND_BINARY = "csound"
+"""Path to csound binary."""
+
 SEQUENTIAL_EVENT_ANNOTATION = ";; NEW SEQUENTIAL EVENT\n;;"
 """Annotation in Csound Score files when a new :class:`SequentialEvent` starts."""